create test directory and move mocks under test directory
